frameworks/fastapi/004_fastapi_celery_redis_flower/proj/requirements.txt
aiodataloader==0.4.0
alembic==1.13.0
amqp==5.2.0
aniso8601==9.0.1
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
asgiref==3.7.2
bcrypt==4.1.1
billiard==4.2.0
black==23.11.0
celery==5.3.6
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
fastapi==0.104.1
flower==2.0.1
graphene==3.3
graphene-sqlalchemy==3.0.0rc1
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.2
h11==0.14.0
humanize==4.9.0
idna==3.6
kombu==5.3.4
Mako==1.3.0
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==4.1.0
prometheus-client==0.19.0
promise==2.3
prompt-toolkit==3.0.41
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
pytz==2023.3.post1
redis==5.0.1
regex==2023.10.3
Rx==3.2.0
singledispatch==4.1.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
starlette-graphene3==0.6.0
toml==0.10.2
tornado==6.4
typing_extensions==4.8.0
tzdata==2023.3
uvicorn==0.24.0.post1
vine==5.1.0
wcwidth==0.2.12
